# microservices/auth-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.BindingRequest:
    properties:
      person_id:
        type: string
      primary:
        type: boolean
      type:
        type: string
    required:
    - person_id
    - type
    type: object
  handler.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handler.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: LLM Filter System Authentication Service API
  title: Auth Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录
      parameters:
      - description: 登录凭证
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 用户名或密码错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户登录
      tags:
      - 认证管理
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 注册信息
        in: body
        name: details
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误或用户已存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户注册
      tags:
      - 认证管理
  /bindings:
    post:
      consumes:
      - application/json
      description: 将当前用户绑定到学生或教师身份
      parameters:
      - description: 绑定信息
        in: body
        name: binding
        required: true
        schema:
          $ref: '#/definitions/handler.BindingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 绑定成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建用户绑定
      tags:
      - 绑定管理
  /bindings/{person_id}:
    delete:
      consumes:
      - application/json
      description: 解除当前用户的指定身份绑定
      parameters:
      - description: 人员ID
        in: path
        name: person_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 解绑成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 解除用户绑定
      tags:
      - 绑定管理
  /bindings/me:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的主身份绑定信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 未找到主绑定
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取我的主绑定
      tags:
      - 绑定管理
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"